requirements.txt
flask
flask-sqlalchemy
flask_script
flask-migrate==2.7.0
flask-cors
jose
python-jose
pymysql
mysql-connector-python
pytest
requests
configparser
minio
pillow
passlib
bcrypt

gkestor-common-logger